requirements.txt
fastapi
uvicorn
python-multipart
PyPDF2
google-generativeai
motor
pydantic-settings